+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 3.0.4 (30 Nov 2020)
 - Fix formatting/ToString of char.
+- Fix issue when using a local variable inside an object initializer
 
 ## 3.0.3 (29 Nov 2020)
 - Fix parsing of comment with a code exit when parsing mode is script only.